Changed HelpIntentHandler to be independent from BaseHandlers
Restructured the IntentHandlers to have HelpIntentHandler be separate for easier editing and possible future work.  Also, fixed a small text issue in the welcome message calling the app "PSU Helper"
